CYBEREASON:: Refacto & split into smaller files + get bulk of max 1000 logs to avoid API error
