+++ vulture_os/toolkit/api_parser/cybereason/malware.py
@@ -1,0 +1,66 @@
+import json
+from datetime import datetime
+
+from pytz import utc
+
+
+class Malware:
+
+    MALWARE_URI = "rest/malware/query"
+
+    def __init__(self, instance: object) -> None:
+        self.instance = instance
+
+    def get_logs(self, since: object, to: object) -> (list, object, bool):
+        malware_uri = f"{self.instance.host}/{self.MALWARE_URI}"
+
+        query = {
+            'filters': [{
+                'fieldName': 'timestamp',
+                'operator': 'GreaterThan',
+                'values': [int(since.timestamp()) * 1000]
+            },{
+                'fieldName': 'timestamp',
+                'operator': 'LessOrEqualsTo',
+                'values': [int(to.timestamp()) * 1000]
+            }],
+            'limit': 1000,
+            'offset': 0,
+            'search': '',
+            'sortDirection': 'DESC',
+            'sortingFieldName': 'timestamp'
+        }
+
+        logs = []
+        offset = 0
+        has_more_logs = True
+        last_timestamp = None
+        # Get bulk of 1000 max logs
+        while len(logs) < 1000:
+            ret = self.instance.execute_query("POST", malware_uri, query)
+            if 'data' not in ret:
+                has_more_logs = False
+                break
+            logs.extend(ret['data']['malwares'])
+            if ret['status'] != 'SUCCESS' or not ret['data']['hasMoreResults']:
+                has_more_logs = False
+                break
+            offset += 1
+            query['offset'] = offset
+        if logs:
+            # Make sure logs are sorted by timestamps
+            # ISO8601 timestamps are sortable as strings
+            last_timestamp = max(x['timestamp'] for x in logs)
+            last_timestamp = datetime.fromtimestamp(int(last_timestamp) / 1000, tz=utc)
+        return logs, last_timestamp, has_more_logs
+
+    def add_enrichment(self, alert: dict) -> dict:
+        alert['id'] = alert['guid']
+        alert['url'] = self.instance.host
+        alert['kind'] = 'malwares'
+        return alert
+
+    @staticmethod
+    def format_log(log: dict) -> json:
+        log['original_timestamp'] = log.pop("timestamp")
+        return json.dumps(log)
